Changing code so that the ASIFT scratch directory doesn't need to be manually set up first, before first run. Read ASIFT Ticket #30 for more details.
This isn't a "user" decision, but more for middleware software managers. When setting up the code to run first time, you'll want to create a Scratch directory where you want the temp-files written, and put that location in file_locations.py. Otherwise, it'll just create a directory where it currently is located in "file_locations.py," which is among the source-code directories. Up to you where you want it.